Revert "Burgershop api sqlite"
